server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-JWT-Extended
Flask-Bcrypt
Pillow
python-dotenv